20.05

+++ Assistant/app/src/main/java/com/diploma/assistant/view/adapter/RecycleViewTasks.java
@@ -1,5 +1,6 @@
 package com.diploma.assistant.view.adapter;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Build;
@@ -23,8 +24,11 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.diploma.assistant.R;
 import com.diploma.assistant.model.entity.resource_service.TaskDto;
+import com.diploma.assistant.model.enumaration.TypeUserEnum;
+import com.diploma.assistant.service.account_manager.AuthenticatorService;
 import com.diploma.assistant.view.ui.sign_up.activity.sign_up_2.CheckStringLine;
 import com.diploma.assistant.view.ui.tasks.admin.TaskActivity;
+import com.diploma.assistant.view.ui.tasks.user.TaskFragment;
 import com.diploma.assistant.view_model.TasksViewModel;
 
 import java.io.Serializable;
@@ -32,8 +36,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
+import javax.crypto.SecretKey;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+
 public class RecycleViewTasks extends RecyclerView.Adapter<RecycleViewTasks.ViewHolderTasks> implements Serializable, Filterable {
-    private final Context context;
+    private final Activity activity;
     private final List<TaskDto> items;
     private final List<TaskDto> itemsFull;
     private final ViewModelStoreOwner viewModelStoreOwner;
@@ -41,8 +52,8 @@ public class RecycleViewTasks extends RecyclerView.Adapter<RecycleViewTasks.View
     private String token;
     private final String name;
 
-    public RecycleViewTasks(Context context, List<TaskDto> items, String id, ViewModelStoreOwner v, LifecycleOwner lifecycle, String name) {
-        this.context = context;
+    public RecycleViewTasks(Activity activity, List<TaskDto> items, String id, ViewModelStoreOwner v, LifecycleOwner lifecycle, String name) {
+        this.activity = activity;
         this.items = items;
         this.viewModelStoreOwner = v;
         this.lifecycle = lifecycle;
@@ -53,7 +64,7 @@ public class RecycleViewTasks extends RecyclerView.Adapter<RecycleViewTasks.View
     @NonNull
     @Override
     public ViewHolderTasks onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        return new ViewHolderTasks(LayoutInflater.from(context).inflate(R.layout.tasks_list_item, parent, false));
+        return new ViewHolderTasks(LayoutInflater.from(activity.getApplicationContext()).inflate(R.layout.tasks_list_item, parent, false));
     }
 
     @RequiresApi(api = Build.VERSION_CODES.Q)
@@ -62,40 +73,54 @@ public class RecycleViewTasks extends RecyclerView.Adapter<RecycleViewTasks.View
         String time = CheckStringLine.parserData(items.get(position).getAddedData());
         holder.datetime.setText(time);
         holder.article.setText(items.get(position).getTitle());
+        AuthenticatorService accounts = new AuthenticatorService(activity);
+        token = accounts.getElementFromSet("Bearer", "jwt_token", "com.assistant.emmotechie.PREFERENCE_FILE_KEY");
 
-        holder.itemView.setOnLongClickListener(v -> {
-            Context wrapper = new ContextThemeWrapper(v.getContext(), R.style.PopMenu);
-            PopupMenu popup = new PopupMenu(wrapper, v);
-            popup.setForceShowIcon(true);
-            popup.inflate(R.menu.tasks_current_menu);
-            popup.setGravity(Gravity.CENTER);
+        SecretKey secret = Keys.hmacShaKeyFor(Decoders.BASE64.decode("uu74l8S6ewO/Nmrh3waPdCfyF7UFTUtFoI44Z5c75X0="));
+        Claims claims = Jwts.parserBuilder().setSigningKey(secret).build().parseClaimsJws(token.replace("Bearer ", "")).getBody();
+        List<String> roles = (List<String>) claims.get("role");
 
-            popup.setOnMenuItemClickListener(item -> {
-                final int itemId = item.getItemId();
-                if (itemId == R.id.delete_tasks) {
-                    TasksViewModel tasksViewModel = new ViewModelProvider(viewModelStoreOwner).get(TasksViewModel.class);
-                    tasksViewModel.deleteTask(token, items.get(position).getId()).observe(lifecycle, l -> {
-                        if(l.equals(true)){
-                            items.remove(position);
-                            notifyItemRemoved(position);
-                            notifyItemRangeChanged(position, items.size());
-                            Toast.makeText(context, "Завдання успішно видалене!", Toast.LENGTH_SHORT).show();
-                        }
-                        else Toast.makeText(context, "Щось пішло не так, спробуйте будь-ласка ще раз", Toast.LENGTH_SHORT).show();
-                    });
-                }
+        if(roles.get(0).equals(TypeUserEnum.ADMIN.getTypeUserName())) {
+            holder.itemView.setOnLongClickListener(v -> {
+                Context wrapper = new ContextThemeWrapper(v.getContext(), R.style.PopMenu);
+                PopupMenu popup = new PopupMenu(wrapper, v);
+                popup.setForceShowIcon(true);
+                popup.inflate(R.menu.tasks_current_menu);
+                popup.setGravity(Gravity.CENTER);
+
+                popup.setOnMenuItemClickListener(item -> {
+                    final int itemId = item.getItemId();
+                    if (itemId == R.id.delete_tasks) {
+                        TasksViewModel tasksViewModel = new ViewModelProvider(viewModelStoreOwner).get(TasksViewModel.class);
+                        tasksViewModel.deleteTask(token, items.get(position).getId()).observe(lifecycle, l -> {
+                            if (l.equals(true)) {
+                                items.remove(position);
+                                notifyItemRemoved(position);
+                                notifyItemRangeChanged(position, items.size());
+                                Toast.makeText(activity.getApplicationContext(), "Завдання успішно видалене!", Toast.LENGTH_SHORT).show();
+                            } else
+                                Toast.makeText(activity.getApplicationContext(), "Щось пішло не так, спробуйте будь-ласка ще раз", Toast.LENGTH_SHORT).show();
+                        });
+                    }
+                    return false;
+                });
+
+                popup.show();
                 return false;
             });
 
-            popup.show();
-            return false;
-        });
-
-        holder.itemView.setOnClickListener(v -> {
-            context.startActivity(new Intent(context, TaskActivity.class)
-                    .putExtra("id", items.get(position).getId())
-                    .putExtra("name", name));
-        });
+            holder.itemView.setOnClickListener(v -> {
+                activity.startActivity(new Intent(activity.getApplicationContext(), TaskActivity.class)
+                        .putExtra("id", items.get(position).getId())
+                        .putExtra("name", name));
+            });
+        } else {
+            holder.itemView.setOnClickListener(v -> {
+                activity.startActivity(new Intent(activity.getApplicationContext(), TaskFragment.class)
+                        .putExtra("id", items.get(position).getId())
+                        .putExtra("name", name));
+            });
+        }
     }
 
     @Override
@@ -146,5 +171,4 @@ public class RecycleViewTasks extends RecyclerView.Adapter<RecycleViewTasks.View
             article = itemView.findViewById(R.id.article_task);
         }
     }
-
 }
